[Fix] slow prototype status retrieval

+++ prototypes/backend/api.py
@@ -11,29 +11,31 @@ app = Flask(__name__)
 ROOT_DIR = "/usr/src/prototypes/generated_prototypes"
 
 manager = Manager()
-running_prototypes = manager.dict()
 lock = Lock()
 
-def stop_prototypes_func():
-    with lock:
-        for prototype_name in list(running_prototypes.keys()):
-            pid, _ = running_prototypes[prototype_name]
-            os.kill(pid, signal.SIGTERM)
-            del running_prototypes[prototype_name]
-    return True
+running_prototype = manager.dict()
 
+
+def stop_prototype():
+    with lock:
+        if "name" in running_prototype:
+            pid = running_prototype["pid"]
+            os.kill(pid, signal.SIGTERM)
+            running_prototype.clear()
 
 
 def start_prototype(prototype_name: str):
     with lock:
-        if prototype_name in running_prototypes:
-            return running_prototypes[prototype_name]
-        
+        if "name" in running_prototype:
+            pid = running_prototype["pid"]
+            os.kill(pid, signal.SIGTERM)
+            running_prototype.clear()
+
         prototype_path = os.path.join(ROOT_DIR, prototype_name)
         if not os.path.isdir(prototype_path):
             return None
 
-        port = 8020 # TODO: put in env
+        port = 8020  # TODO: put in env
         process = subprocess.Popen(
             ["python", "manage.py", "runserver", f"0.0.0.0:{port}"],
             cwd=prototype_path,
@@ -43,43 +45,44 @@ def start_prototype(prototype_name: str):
         time.sleep(2)
         if process.poll() is not None:
             return None
-    
-        running_prototypes[prototype_name] = (process.pid, port)
+
+        running_prototype["name"] = prototype_name
+        running_prototype["pid"] = process.pid
+        running_prototype["port"] = port
         return process, port
 
 
 @app.route('/run/<prototype_name>', methods=['POST'])
 def run_prototype(prototype_name: str):
-    stop_prototypes_func()
+    stop_prototype()
     result = start_prototype(prototype_name)
     if result:
-        pid, port = result
-        return redirect(f"http://prototype.ai4mde.localhost", code=307)
+        _, port = result
+        return redirect(f"http://prototype.ai4mde.localhost:{port}", code=307) # TODO: put in env
     else:
         abort(404)
 
 
 @app.route('/stop_prototypes', methods=['POST'])
 def stop_prototypes():
-    stop_prototypes_func()
+    stop_prototype()
     return f"Stopped all running prototypes", 200
 
 
-@app.route('/status/<prototype_name>', methods=['GET'])
-def status_prototype(prototype_name: str):
+@app.route('/active_prototype', methods=['GET'])
+def get_active_prototype():
     with lock:
-        if prototype_name in running_prototypes:
-            pid, port = running_prototypes[prototype_name]
+        if "name" in running_prototype:
             return {
-                "prototype_name": prototype_name,
+                "prototype_name": running_prototype["name"],
                 "running": True,
-                "pid": pid,
+                "pid": running_prototype["pid"],
                 "ip": socket.gethostbyname(socket.gethostname()),
-                "port": port
+                "port": running_prototype["port"]
             }
         else:
             return {
-                "prototype_name": prototype_name,
+                "prototype_name": None,
                 "running": False,
                 "pid": None,
                 "ip": None,
@@ -98,7 +101,7 @@ def generate_prototype():
         subprocess.run([GENERATOR_PATH, prototype_name, metadata], check=True)
     except subprocess.CalledProcessError:
         return f"Failed to generate {prototype_name} prototype", 500
-    
+
     if 'database_prototype_name' in data:
         database_prototype_name = data.get('database_prototype_name')
         try:
@@ -110,13 +113,11 @@ def generate_prototype():
 
 @app.route('/remove', methods=['DELETE'])
 def remove_prototype():
-    REMOVER_PATH = "/usr/src/prototypes/backend/generation/remover.sh"
+    REMOVER_PATH = "/usr/src/prototypes/backend/generation/remover.sh" # TODO: put in env
     data = request.json
     prototype_name = data.get('prototype_name')
-    if prototype_name in running_prototypes:
-        pid, _ = running_prototypes[prototype_name]
-        os.kill(pid, signal.SIGTERM)
-        del running_prototypes[prototype_name]
+    if "name" in running_prototype and running_prototype["name"] == prototype_name:
+        stop_prototype()
     try:
         subprocess.run([REMOVER_PATH, prototype_name], check=True)
     except subprocess.CalledProcessError:
@@ -125,4 +126,4 @@ def remove_prototype():
 
 
 if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=8010, debug=True)
+    app.run(host='0.0.0.0', port=8010, debug=True) # TODO: put in env
